Vertex Translation Partial Implementation
I have partially implemented the ability to translate vertices by clicking and dragging on the axis frame that appears when a node is selected, although it isn't entirely stable. The program will crash if a node attempts to "go past" (i.e. exceed the Z or X position) one of its neighboring vertices. I've notified the professor about how to fix this.